windicate.py 0.0.3: add two-way sync between indications and weechat

+++ python/windicate.py
@@ -1,40 +1,58 @@
 # Author: Leon Bogaert <leon AT tim-online DOT nl>
+# Author: Stacey Sheldon <stac AT solidgoldbomb DOT org>
 # This Plugin Calls the libindicate bindings via python when somebody says your
 # nickname, sends you a query, etc.
-# To make it work, you may need to download: python-indicate and python-dbus
+# To make it work, you may need to download:
+#    python-indicate
+#    python-dbus
+#    wmctrl
 # Requires Weechat 0.3.0
 # Released under GNU GPL v2
 #
 # 2010-09-22, Leon <leon@tim-online.nl>:
 #     version 0.0.1 Intial release
-# 
-# @TODO: find out how to jump to buffer/line
-# @TODO: how to communicate the click to weechat
+# 2013-04-14, Stacey Sheldon <stac@solidgoldbomb.org>
+#     version 0.0.2 Added two-way sync between indications and weechat
+# 2013-05-01, Stacey Sheldon <stac@solidgoldbomb.org>
+#     version 0.0.3 More graceful handling of missing dependencies
+#
 # @TODO: decide what to do if a user clicks an indicator an then start typing:
 #        * leave indicators alone
 #        * remove indicators in the "neighbourhood"
-#        * If a user cliks indicator: indactor dissapears
 #        * On click group: remove all indicators
-#        * When visiting buffer: remove indicators
 
 import dbus.service
-import inspect
 import os
-import tempfile
+
+SCRIPT_NAME    = "windicate"
+SCRIPT_AUTHOR  = "Leon Bogaert"
+SCRIPT_VERSION = "0.0.3"
+SCRIPT_LICENSE = "GPL"
+SCRIPT_DESC    = "fills the indicate applet"
+
+DBUS_CONNECTION = 'org.weechat.scripts.windicate'
+DBUS_OBJ_PATH   = '/org/weechat/scripts/windicate'
 
 class DBUSService(dbus.service.Object):
     def __init__(self, messageMenu):
-        bus_name = dbus.service.BusName('org.weechat.scripts.windicate', bus=dbus.SessionBus())
-        dbus.service.Object.__init__(self, bus_name, '/org/weechat/scripts/windicate')
+        bus_name = dbus.service.BusName(DBUS_CONNECTION, bus=dbus.SessionBus())
+        dbus.service.Object.__init__(self, bus_name, DBUS_OBJ_PATH)
         self.messageMenu = messageMenu
- 
+
     @dbus.service.method('org.weechat.scripts.windicate')
-    def add_message(self, channel, sender, body):
-        return self.messageMenu.add_message(channel, sender, body)
+    def add_message(self, buffer, brief, sender, body):
+        return self.messageMenu.add_message(buffer, brief, sender, body)
+
+    @dbus.service.method('org.weechat.scripts.windicate')
+    def del_messages(self, buffer):
+        return self.messageMenu.del_messages(buffer)
 
 class MessageMenu(object):
-    def __init__(self):
+    def __init__(self, weechat_fifo, weechat_windowid):
         self.messages = []
+
+        self.weechat_fifo = weechat_fifo
+        self.weechat_windowid = weechat_windowid
 
         server = pyindicate.indicate_server_ref_default()
         server.set_type("message.im")
@@ -43,13 +61,27 @@ class MessageMenu(object):
         server.show()
 
     def desktop_file(self):
+        DESKTOP_ENTRY = """\
+[Desktop Entry]
+Encoding=UTF-8
+MultipleArgs=false
+Terminal=true
+Exec=weechat-curses
+Icon=weechat
+Type=Application
+Categories=Network;IRCClient;
+StartupNotify=false
+Name=Weechat
+GenericName=IRC Client
+"""
         file = "/usr/share/applications/weechat.desktop"
 
         if os.path.isfile(file):
             return file
 
+        import tempfile
         f = tempfile.NamedTemporaryFile(suffix='indicator', delete=False)
-        f.write("[Desktop Entry]\nEncoding=UTF-8\nMultipleArgs=false\nTerminal=true\nExec=weechat-curses\nIcon=weechat\nType=Application\nCategories=Network;IRCClient;\nStartupNotify=false\nName=Weechat\nGenericName=IRC Client")
+        f.write(DESKTOP_ENTRY)
         f.close()
         self.tmp_file = f
 
@@ -59,18 +91,44 @@ class MessageMenu(object):
         if self.tmp_file and os.file.exists(self.tmp_file):
             os.unlink(self.tmp_file)
 
-    def server_click(self, server, time):
-        print "Server clicked!"
+    def raise_window(self):
+        import subprocess
+        try:
+            retcode = subprocess.call('wmctrl -i -a %s' % self.weechat_windowid, shell=True)
+            if retcode < 0:
+                # print >>sys.stderr, "Child was terminated by signal", -retcode
+                pass
+            else:
+                # print >>sys.stderr, "Child returned", retcode
+                pass
+        except OSError as e:
+            # print >>sys.stderr, "Execution failed:", e
+            pass
 
-    def add_message(self, channel, sender, body):
+    def server_click(self, server, time):
+        # tell weechat to select the first active buffer
+        with open (self.weechat_fifo, 'a') as f:
+            f.write ("irc.server.freenode */input jump_smart\n")
+        # raise the weechat window to the foreground
+        self.raise_window()
+
+    def add_message(self, buffer, brief, sender, body):
         for message in self.messages:
-            if message.channel == channel and message.sender == sender:
+            if message.buffer == buffer and message.sender == sender:
                 return message.update_time()
 
-        return self.messages.append(Message(self, channel, sender, body))
+        return self.messages.append(Message(self, buffer, brief, sender, body))
+
+    def del_messages(self, buffer):
+        # remove all pending indications for this buffer
+        for message in self.messages:
+            if message.buffer == buffer:
+                self.messages.remove(message)
+                message.indicator.hide()
+        return True
 
 class Message(object):
-    def __init__(self, mm, channel, sender, message):
+    def __init__(self, mm, buffer, brief, sender, message):
         # Setup the message
         try:
             # Ubuntu 9.10 and above
@@ -80,7 +138,7 @@ class Message(object):
             indicator = pyindicate.IndicatorMessage()
 
         indicator.set_property("subtype", "im")
-        indicator.set_property("sender", "%s (%s)" % (sender, channel))
+        indicator.set_property("sender", "%s (%s)" % (sender, brief))
         indicator.set_property("body", message)
         indicator.set_property_time("time", time())
         indicator.set_property('draw-attention', 'true');
@@ -90,22 +148,26 @@ class Message(object):
         self.indicator = indicator
         self.sender = sender
         self.message = message
-        self.channel = channel
+        self.brief = brief
         self.messageMenu = mm
+        self.buffer = buffer
 
     def update_time(self):
         self.indicator.set_property_time("time", time())
-        
+
     def message_clicked(self, indicator, time):
-        #How can I make weechat go there?? (/buffer self.channeli)
-        #Maybe if I can get dbus running in a weechat script...
         self.messageMenu.messages.remove(self)
         indicator.hide()
+        # tell weechat to select the buffer that triggered this indication
+        with open (self.messageMenu.weechat_fifo, 'a') as f:
+            f.write ("core.weechat */buffer %s\n" % self.buffer)
+        # raise the weechat window to the foreground
+        self.messageMenu.raise_window()
 
 class WindicateServer(object):
-    def __init__(self):
+    def __init__(self, weechat_fifo, weechat_windowid):
         DBusGMainLoop(set_as_default=True)
-        mm = MessageMenu()
+        mm = MessageMenu(weechat_fifo, weechat_windowid)
         dbs = DBUSService(mm)
 
         loop = gobject.MainLoop()
@@ -115,16 +177,21 @@ class Subprocess(object):
     p = None
 
     @classmethod
-    def start(cls):
-        file = inspect.getfile( inspect.currentframe())
+    def start(cls, fifo_filename, window_windowid):
+        import inspect
+        file = inspect.getfile(inspect.currentframe())
 
         import subprocess
-        args = ["/usr/bin/python", file]
+        args = ["/usr/bin/python",
+                file,
+                fifo_filename,
+                weechat_windowid]
         cls.p = subprocess.Popen(args)
 
     @classmethod
     def stop(cls):
-        cls.p.terminate()
+        if cls.p != None:
+            cls.p.terminate()
 
 try:
     import weechat
@@ -137,12 +204,12 @@ except ImportError:
     import dbus
     from dbus.mainloop.glib import DBusGMainLoop
     from time import time
-    ws = WindicateServer()
+    import sys # argv
+
+    # arguments: weechat_fifo_filename weechat_windowid
+    ws = WindicateServer(sys.argv[1], sys.argv[2])    # doesn't return
 
 ##### FUNCTIONS #####
-def windicate_server_ended(data, command, rc, stdout, stderr):
-    print "windicate_server_ended"
-    return weechat.WEECHAT_RC_OK
 
 def weechat_script_end():
     Subprocess.stop()
@@ -151,69 +218,90 @@ def weechat_script_end():
 def notify_msg(data, bufferp, time, tags, display, is_hilight, prefix, msg):
     """Sends highlighted message to be printed on notification"""
 
-    if ('notify_private' in tags and 
+    if ('notify_private' in tags and
         weechat.config_get_plugin('show_priv_msg') == "on") \
         or (is_hilight == "1" and \
         weechat.config_get_plugin('show_hilights') == "on"):
 
-        if not weechat.buffer_get_string(bufferp, "short_name"):
-            buffer = weechat.buffer_get_string(bufferp, "name")
+        # grab the fully qualified buffer name so we can jump to it later
+        buffer = weechat.buffer_get_string(bufferp, "name")
+
+        # choose an appropriate brief name to display in the indicator applet
+        if 'notify_private' in tags:
+            brief = "private"
         else:
-            buffer = weechat.buffer_get_string(bufferp, "short_name")
-
-        if ('notify_private' in tags):
-            buffer = "private"
-
-        add_message(buffer, prefix, msg)
+            # prefer short_name
+            brief = weechat.buffer_get_string(bufferp, "short_name")
+            if not brief:
+                # fall back to full name
+                brief = buffer
 
         if weechat.config_get_plugin('debug') == "on":
-            print prefix
+            print "buffer: " + buffer
+            print "brief: " + brief
+            print "prefix: " + prefix
+            print "msg: " + msg
+
+        # Create an object that will proxy for a particular remote object.
+        bus = dbus.SessionBus()
+        remote_object = bus.get_object(DBUS_CONNECTION, DBUS_OBJ_PATH)
+        remote_object.add_message(buffer, brief, prefix, msg)
 
     return weechat.WEECHAT_RC_OK
 
-def notify_show_hi(data, signal, message):
-    print "data: " + data
-    print "signal: " + data
-    print "message: " + message
-    return weechat.WEECHAT_RC_OK
-
-def notify_show_priv(data, signal, message):
-    print "data: " + data
-    print "signal: " + data
-    print "message: " + message
-    return weechat.WEECHAT_RC_OK
-
-def add_message(channel, sender, body):
-    bus = dbus.SessionBus()
+def buffer_switched(data, signal, signal_data):
+    buffer = weechat.buffer_get_string(signal_data, "name")
+    if weechat.config_get_plugin('debug') == "on":
+        print "data: " + data
+        print "signal: " + signal
+        print "message: " + signal_data
+        print "buffer: " + buffer
 
     # Create an object that will proxy for a particular remote object.
-    remote_object = bus.get_object("org.weechat.scripts.windicate", # Connection name
-                                   "/org/weechat/scripts/windicate" # Object's path
-                                  )
-    remote_object.add_message(channel, sender, body)
+    bus = dbus.SessionBus()
+    remote_object = bus.get_object(DBUS_CONNECTION, DBUS_OBJ_PATH)
+    remote_object.del_messages(buffer)
+
+    return weechat.WEECHAT_RC_OK
+
 ##### END FUNCTIONS #####
 
-weechat.register("windicate", "Leon Bogaert", "0.0.1", "GPL",
-                 "fills the indicate applet", "weechat_script_end", "")
-
-# script options
 settings = {
-    "show_hilights" : "on",
-    "show_priv_msg" : "on",
-    "time_between_msg" : "5",
+    'show_hilights' : ('on', 'Should hilights trigger indications' ),
+    'show_priv_msg' : ('on', 'Should privmsgs trigger indications' ),
 }
 
-# Init everything
-for option, default_value in settings.items():
-    if weechat.config_get_plugin(option) == "":
-        weechat.config_set_plugin(option, default_value)
+if weechat.register(SCRIPT_NAME,
+                    SCRIPT_AUTHOR,
+                    SCRIPT_VERSION,
+                    SCRIPT_LICENSE,
+                    SCRIPT_DESC,
+                    "weechat_script_end",
+                    ""):
+    version = weechat.info_get('version_number', '') or 0
 
-Subprocess.start()
+    # Init everything
+    for option, default_desc in settings.items():
+        if not weechat.config_is_set_plugin(option):
+            weechat.config_set_plugin(option, default_desc[0])
+        if int(version) >= 0x00030500:
+            weechat.config_set_desc_plugin(option, default_desc[1])
 
-#weechat.hook_process("/usr/bin/python %s" % (file,),
-                     #-1, "windicate_server_ended", "")
+    # Perform some sanity checks to make sure we have everything we need to run
+    sanity = True
+    weechat_windowid = os.environ.get('WINDOWID')
+    if weechat_windowid == None:
+        weechat.prnt("", "%sEnvironment variable WINDOWID not set.  This script requires an X environment to run." % weechat.prefix("error"))
+        sanity = False
 
-# Hook privmsg/hilights
-weechat.hook_print("", "", "", 1, "notify_msg", "")
-#weechat.hook_signal("weechat_highlight", "notify_show_hi", "")
-#weechat.hook_signal("weechat_pv", "notify_show_priv", "")
+    fifo_filename = weechat.info_get("fifo_filename", "")
+    if fifo_filename == "":
+        weechat.prnt("", "%sWeechat variable fifo_filename is not set.  Is the fifo plugin enabled?" % weechat.prefix("error"))
+        sanity = False
+
+    if sanity:
+        Subprocess.start(fifo_filename, weechat_windowid)
+
+        # Hook privmsg/hilights
+        weechat.hook_print("", "", "", 1, "notify_msg", "")
+        weechat.hook_signal("buffer_switch", "buffer_switched", "")
